[REFACTOR] Comment 삭제 및 수정 유저 확인 기능 서비스로 이동

+++ src/main/java/techcourse/myblog/service/CommentService.java
@@ -3,9 +3,11 @@ package techcourse.myblog.service;
 import org.springframework.stereotype.Service;
 import techcourse.myblog.domain.*;
 import techcourse.myblog.service.dto.CommentRequest;
+import techcourse.myblog.service.dto.UserResponse;
 import techcourse.myblog.service.exception.CommentNotFoundException;
 import techcourse.myblog.service.exception.NoArticleException;
 import techcourse.myblog.service.exception.NoUserException;
+import techcourse.myblog.service.exception.NotSameAuthorException;
 
 import javax.transaction.Transactional;
 import java.util.List;
@@ -40,12 +42,28 @@ public class CommentService {
         return commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException("존재하지 않는 댓글입니다."));
     }
 
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, UserResponse userResponse) {
+        Comment comment = findCommentById(commentId);
+        User author = userRepository.findUserByEmail(userResponse.getEmail()).orElseThrow(() -> new NoUserException("유저가 존재하지 않습니다."));
+
+        checkSameAuthor(comment, author);
+
         commentRepository.deleteById(commentId);
     }
 
     @Transactional
-    public void updateComment(Comment comment, CommentRequest commentRequest) {
+    public void updateComment(Long commentId, UserResponse userResponse, CommentRequest commentRequest) {
+        Comment comment = findCommentById(commentId);
+        User author = userRepository.findUserByEmail(userResponse.getEmail()).orElseThrow(() -> new NoUserException("유저가 존재하지 않습니다."));
+
+        checkSameAuthor(comment, author);
+
         comment.changeContents(commentRequest.getContents());
+    }
+
+    private void checkSameAuthor(Comment comment, User author) {
+        if (!comment.isSameAuthor(author)) {
+            throw new NotSameAuthorException("해당 댓글의 작성자가 아닙니다.");
+        }
     }
 }
